unsigned long should not be used to store pointers
On LLP64 systems (read Windows) unsigned long is only 4 bytes
wide, which makes it very unsuitable for storing pointers.

+++ src/server/ns_turn_maps.h
@@ -45,7 +45,7 @@ typedef struct _ur_map ur_map;
 //////////////// Common Definitions //////
 
 typedef uint64_t ur_map_key_type;
-typedef unsigned long ur_map_value_type;
+typedef uintptr_t ur_map_value_type;
 
 typedef void (*ur_map_del_func)(ur_map_value_type);
 
@@ -158,7 +158,7 @@ int lm_map_foreach_arg(lm_map* map, foreachcb_arg_type func, void* arg);
 
 //////////////// UR ADDR MAP //////////////////
 
-typedef unsigned long ur_addr_map_value_type;
+typedef uintptr_t ur_addr_map_value_type;
 
 #define ADDR_MAP_SIZE (1024)
 #define ADDR_ARRAY_SIZE (4)
